+++ test/logstop_test.rb
@@ -58,9 +58,9 @@ class LogstopTest < Minitest::Test
     assert_equal "[Ruby] begin [FILTERED] end\n", str.string
   end
 
-  def test_rules
-    assert_filtered "hello", rules: [/hello/, /goodbye/]
-    assert_filtered "goodbye", rules: [/hello/, /goodbye/]
+  def test_extra_rules
+    assert_filtered "hello", extra_rules: [/hello/, /goodbye/]
+    assert_filtered "goodbye", extra_rules: [/hello/, /goodbye/]
   end
 
   private